% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_perm.R
\name{MAP_cluster}
\alias{MAP_cluster}
\title{return the maximum-a-posteriori cluster label for an individual given K and rep}
\usage{
MAP_cluster(df)
}
\arguments{
\item{df}{long-format data frame like that which comes out of \code{\link{read_results}}.
Minimally it should have the columns \code{K}, \code{rep}, \code{Index}, \code{cluster}, and
\code{probability}.}
}
\value{
a summarised data frame with the column \code{map_cluster} in place of \code{cluster}.
}
\description{
return the maximum-a-posteriori cluster label for an individual given K and rep
}

