% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_read_and_process.R
\name{read_and_process_structure_output}
\alias{read_and_process_structure_output}
\title{read and process all the structure output files}
\usage{
read_and_process_structure_output(Dir)
}
\arguments{
\item{Dir}{directory in which the output resides}
}
\value{
This returns a list with two components: \code{q} is a data frame
of inferred ancestry proportions. \code{traces} is a data frame of the
traces of a few key variables.
}
\description{
reads in the structure output files and the trace files and
does relabeling of clusters to make them as concordant as possible.
}

