% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slurp_output.R
\name{read_results}
\alias{read_results}
\title{read all the results in from all the files in the specified output directory}
\usage{
read_results(D)
}
\arguments{
\item{D}{the directory holding all the text output from the structure_run() function}
}
\description{
this function also puts everything into long (tidy) format
}

