% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurp_output.R
\name{read_traces}
\alias{read_traces}
\title{read all the traces in from all the files in the specified output directory}
\usage{
read_traces(D)
}
\arguments{
\item{dir}{the directory holding all the text output from the structure_run() function}
}
\description{
this function also puts everything into long (tidy) format
}

