% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/label_perm.R
\name{relabel_map_clusters}
\alias{relabel_map_clusters}
\title{for each rep > 1 within a given K, return the permutation index (from kperm) that maximizes the
concordance in map_cluster between the first rep and the current one.}
\usage{
relabel_map_clusters(df)
}
\arguments{
\item{df}{this should be the output from MAP_cluster}
}
\description{
for each rep > 1 within a given K, return the permutation index (from kperm) that maximizes the
concordance in map_cluster between the first rep and the current one.
}

