% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slurp_output.R
\name{slurp_traces}
\alias{slurp_traces}
\title{slurp out the traces from a file and return them as a data frame with headers}
\usage{
slurp_traces(file, K, rep)
}
\arguments{
\item{file}{the name of the stdout file to slurp}

\item{K}{the value of K that this was run at}

\item{rep}{Which rep number was this structure run.}
}
\description{
slurp out the traces from a file and return them as a data frame with headers
}

