% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_struct.R
\name{struct_binary_path}
\alias{struct_binary_path}
\title{check what type of OS and then figure out the path to the structure executable appropriately}
\usage{
struct_binary_path()
}
\description{
check what type of OS and then figure out the path to the structure executable appropriately
}

