% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_struct.R
\name{structure_runs}
\alias{structure_runs}
\title{run structure at a specified set of K's for a specified number of reps}
\usage{
structure_runs(outdir, K, reps, seed)
}
\arguments{
\item{outdir}{The name of the output directory to use.  The directory will be
created if it does not already exist.  Note, if you give a directory that already
has output, that old output might get overwritten.}

\item{K}{a vector of K values to simulate at.  They can't be duplicated}

\item{reps}{a vector of the number of reps to simulate for each K value.  This
recycles to the length of the K values vector.}
}
\description{
run structure at a specified set of K's for a specified number of reps
}

